Fix awscli setup
The travis build for BridgeResearcherUI was setting the TRAVIS_TAG
environment variable[1] even though the build was not for a tag.
This made setup_aws_cli.sh[2] skip double interprolation of env vars
causing role_arn[3] to not be set which made the BridgeResearchUI build
to fail.

What we want double interpolation when:
(TRAVIS_TAG is not set OR TRAVIS_TAG is empty) AND
TRAVIS_BRANCH is not master

[1] https://travis-ci.org/Sage-Bionetworks/BridgeResearcherUI/builds/460865390#L554
[2] https://github.com/Sage-Bionetworks/infra-utils/blob/master/ci/setup_aws_cli.sh
[3] https://travis-ci.org/Sage-Bionetworks/BridgeResearcherUI/builds/460865390#L629