Update lint stack name
Instead of getting only the stacks that have been committed to
git repo we want to get all stacks directly from cloudformation.
This will give us the true set of existing stack names to compare
with the new stack name that's being added.

We no longer have to pass in a parameter to the script and we
no longer need to checkout a previous commit which simplifies things.